MDEXP-628 - abstract strategy
